Mostly implement chess (whoops I didn't commit for a while)
The game logic is almost entirely there.
still not done:
- en passant is half implemented
- you can only promote pawns to queens

there's a lot of refactoring i'd like to do and things i'd like to make better as well.
Big thing is that I'm representing potential moves as squares instead of using the Ply type, which requires extra logic to tell if i a move is, eg, a castle.