Refactor to use Ply instead of Square to represent moves everywhere
This compiles but fails to start due to "RangeError: Maximum call stack size exceeded"
I'm committing it so i can come back later and see exactly what the code that produced this error was.
I'm pretty sure I've got an infinite recursion, and it probably has to do with checking for check.