Fix check-checking crash
I added a flag to the function that generates all possible moves to tell it whether to include castling or not.
When checking for check by generating all the possible moves, if we include castling in the set, we cause infinite recursion.
You can't capture by castling so we dont need to include it.
This is absolutely not the most elegant way of solving this problem.