Revamp color usage
Using overlay for possible moves to show a transparent circle over the square instead of making its
background red, which was overly aggressive.
I've also switched to using the same focus color for both white and black, and tweaked the palette
a bit to make this work better.
i think the palette needs further tweaking to look nicer, but this is workable for now.