Change parsing strategy
My previous plan was to validate every ply as we parsed it, and build up a
position as we go. However, I think it makes more sense to do an initial
parsing pass, and collect all the information we're given, and then
use all of those to build up the history.
I think keeping game logic and parsing separate will make life a lot easier.

So i made a new type, PgnPly, which will hold just the information we're
given in the text we're parsing (eg no team info, usually no start square).