Flawed approach to parsing
I realize now that I had a fundamental misunderstanding of how parsers work,
so this doesn't really work, but i wanted to commit it before pivoting so that
i'd have it in history.

i thought that oneOf used the first parser to succeed, but instead it uses the first parser that
chomps anything, which is a big difference.
I have to either add backtracking or do more of a character by character approach.