Parse non pawn/non castling moves
Things like Qf4, Nf3, etc
The tricky part here is that we can optionally get additional
info about the starting position of the piece.
like Ngf3 to specify the knight in the g file.

I basically enumerated all the possibilities and didn't do anything fancy
here. It looks kinda ugly, and I hope I will find a classier way to do it,
but if not that's ok.
Basically, I assume that this plies look like a piece kind followed by up to five
characters. Each of these characters is either a rank, file, or 'x'.
Once I get all of those, I use pattern matching on the list of parsed things to
see what kind of ply it is, and if it doesn't fit any of the patterns that i allow,
reject it.