Automatically test PGN parsing as games are played
I have the app generate the text for the current position, then attempt to parse the
position from that. if the positions dont match, or there is a parsing error,
text is displayed below the board.
Additional status info is printed to the console.
This helped me find an error in knight move parsing pretty quickly.