Prevent some unnecessary calculations when checking for castling
Original version of the function checked if all the squares in between the king
and rook starting positions were threatened before checking if the pieces were
even still in those positions. This didn't case any bugs, but it led to a branch
being hit that I never expected to be hit, so i saw some console messages.
It's also a lot of unnecessary computation, so this might make some stuff faster.