Add a crappy AI
it's just based on some heuristics that i made up on the spot without planning.
This was mostly just fun to play with, but it does allow me to get to some
interesting positions for testing.
It is quite slow: ~5ish seconds per move on my macbook, tens of seconds per move on my phone
And it makes some silly moves. I'm happy with how often it makes sensible looking moves
though, although generally it makes them for the wrong reasons.
This is not a task that elm is well suited for, i think it will be fun to extend this
with some graph search stuff and really try to push the limits and play with
optimization.