Indicate if ply is invalid due to exposing check in error message
This isn't strictly necessary, but it's handy for quickly realizing what
went wrong.
I feel like I might be abusing the result class a bit here, like i could
just use a Maybe String instead, but I feel like this communicates better.
I think it might be a hint that i'm doing this check in the wrong place,
and I should have a custom type for my parsing errors, but this works
for now.