Add input text box for parsing positions from PGN
When a valid position is put in the box, the current board
will be changed to that position.

this can be used interactively instead of clicking or for
pasting in old games.
the parsing is a bit lax, so you can skip move numbers
when using interactively.
eg you can type "e4 e5 Nf3" instead of "1. e4 e5 2. Nf3"